Re-namespace models and extractors
